+++ src/components/paperViewer/DetailCardModal.jsx
