+++ src/pages/PaperListPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import Carousel from '@components/ListPage/Carousel';
+import Carousel from '@components/list/Carousel';
 import styled from 'styled-components';
 import Button from '@components/common/button/Button';
 import recipientsAPI from '@/api/recipientsAPI';
